requirements.txt
alembic==1.13.3
aniso8601==9.0.1
apispec==6.7.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.7.0
click==8.1.7
email_validator==2.2.0
cloud-sql-python-connector[pg8000]
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
GeoAlchemy2==0.16.0
google-cloud-storage
greenlet==3.1.1
gunicorn==21.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging>=21.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
Authlib==1.2.1
requests==2.31.0
Flask-Babel==4.0.0
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.2
WTForms==3.1.2
zipp==3.18.1
